arrow-glacier refactored
